+++ src/components/domain/Page/molecules/PagePreviewCard/PagePreviewCard.tsx
@@ -3,9 +3,9 @@ import { VFC } from 'react';
 import styled from 'styled-components';
 
 import { imagePath } from '~/libs/const/imagePath';
+import { BootstrapBreakpoints } from '~/libs/interfaces/variables';
 
 import { Page } from '~/domains/Page';
-import { BootstrapBreakpoints } from '~/interfaces/variables';
 
 const MAX_WORD_COUNT_OF_BODY = 40;
