+++ src/libs/interfaces/googleAnalytics.ts
