+++ src/libs/interfaces/news.ts
