+++ src/libs/interfaces/paginationResult.ts
