+++ src/libs/interfaces/variables.ts
