+++ src/libs/interfaces/webevNextPage.ts
