+++ src/pages/index.tsx
@@ -6,10 +6,12 @@ import axios from 'axios';
 
 import { WebevOgpHead } from '~/components/common/WebevOgpHead';
 
-import { useLocale } from '~/hooks/useLocale';
 import { imagePath } from '~/libs/const/imagePath';
+import { WebevNextPage } from '~/libs/interfaces/webevNextPage';
+
+import { useLocale } from '~/hooks/useLocale';
+
 import { PageCountupCard } from '~/components/domain/Page/atoms/PageCountupCard';
-import { WebevNextPage } from '~/interfaces/webevNextPage';
 import { DefaultLayout } from '~/components/common/Layout/DefaultLayout';
 
 type Props = {
