+++ src/pages/news/[id].tsx
@@ -6,10 +6,10 @@ import styled from 'styled-components';
 
 import { format } from 'date-fns';
 
-import { News } from '~/interfaces/news';
+import { News } from '~/libs/interfaces/news';
+import { WebevNextPage } from '~/libs/interfaces/webevNextPage';
 import { useLocale } from '~/hooks/useLocale';
 import { WebevOgpHead } from '~/components/common/WebevOgpHead';
-import { WebevNextPage } from '~/interfaces/webevNextPage';
 import { DefaultLayout } from '~/components/common/Layout/DefaultLayout';
 
 type Props = {
