+++ src/pages/news/index.tsx
@@ -5,9 +5,9 @@ import axios from 'axios';
 import styled from 'styled-components';
 
 import { useLocale } from '~/hooks/useLocale';
-import { News } from '~/interfaces/news';
+import { News } from '~/libs/interfaces/news';
+import { WebevNextPage } from '~/libs/interfaces/webevNextPage';
 import { WebevOgpHead } from '~/components/common/WebevOgpHead';
-import { WebevNextPage } from '~/interfaces/webevNextPage';
 import { DefaultLayout } from '~/components/common/Layout/DefaultLayout';
 
 type Props = {
