+++ src/pages/page/[id].tsx
@@ -9,7 +9,7 @@ import { LoginRequiredWrapper } from '~/components/common/Authentication/LoginRe
 import { DashBoardLayout } from '~/components/common/Layout/DashBoardLayout';
 
 import { usePageByPageId } from '~/stores/page';
-import { WebevNextPage } from '~/interfaces/webevNextPage';
+import { WebevNextPage } from '~/libs/interfaces/webevNextPage';
 import { useLocale } from '~/hooks/useLocale';
 
 const Page: WebevNextPage = () => {
